removed more commands file because it is not needed
