--- src/headers/moreCommands.h
@@ -1,17 +1,0 @@
-// CREATED A NEW FILE (moreCommands) BECAUSE FOR SOME REASON THE COMPILER GOES HAYWIRE IF I HAVE EVEN ONE MORE FUNCTION IN THE OTHER FILE (shellCommands)
-#pragma once
-
-
-#include "std.h"
-#include "virtualMemory.h"
-#include "processManager.h"
-#include "screen.h"
-#include "PCI.h"
-#include "memoryAllocator.h"
-#include "demoPrograms.h"
-
-
-
-void cmd_curwide(char** args, unsigned int argCount);
-
-
